# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCAT Coach",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/anki/sync": {
      "post": {
        "tags": [
          "anki"
        ],
        "summary": "Sync Anki",
        "description": "Trigger a one-off Anki deck sync against the configured deck.",
        "operationId": "sync_anki_api_v1_anki_sync_post",
        "parameters": [
          {
            "name": "x-coach-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coach-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sync Anki Api V1 Anki Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anki/review-queue": {
      "get": {
        "tags": [
          "anki"
        ],
        "summary": "Get Review Queue",
        "operationId": "get_review_queue_api_v1_anki_review_queue_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-coach-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coach-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnkiReviewQueueCardOut"
                  },
                  "title": "Response Get Review Queue Api V1 Anki Review Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anki/cards/by-qid/{qid}": {
      "get": {
        "tags": [
          "anki"
        ],
        "summary": "Get Cards For Qid",
        "operationId": "get_cards_for_qid_api_v1_anki_cards_by_qid__qid__get",
        "parameters": [
          {
            "name": "qid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Qid"
            }
          },
          {
            "name": "x-coach-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coach-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnkiCardOut"
                  },
                  "title": "Response Get Cards For Qid Api V1 Anki Cards By Qid  Qid  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnkiCardOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "anki_card_id": {
            "type": "integer",
            "title": "Anki Card Id"
          },
          "deck_name": {
            "type": "string",
            "title": "Deck Name"
          },
          "note_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Id"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "fields_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields Json"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "interval_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Days"
          },
          "ease": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ease"
          },
          "lapses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lapses"
          },
          "queue": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue"
          },
          "sync_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sync At"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/AnkiCardTagOut"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "anki_card_id",
          "deck_name",
          "sync_at"
        ],
        "title": "AnkiCardOut"
      },
      "AnkiCardTagOut": {
        "properties": {
          "tag_raw": {
            "type": "string",
            "title": "Tag Raw"
          },
          "parsed_kind": {
            "type": "string",
            "title": "Parsed Kind"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "question_qid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Qid"
          }
        },
        "type": "object",
        "required": [
          "tag_raw",
          "parsed_kind"
        ],
        "title": "AnkiCardTagOut"
      },
...
      "AnkiReviewQueueCardOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "anki_card_id": {
            "type": "integer",
            "title": "Anki Card Id"
          },
          "deck_name": {
            "type": "string",
            "title": "Deck Name"
          },
          "note_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Id"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "fields_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields Json"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "interval_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Days"
          },
          "ease": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ease"
          },
          "lapses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lapses"
          },
          "queue": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue"
          },
          "sync_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sync At"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/AnkiCardTagOut"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "retention": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retention"
          },
          "retrievability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrievability"
          }
        },
        "type": "object",
        "required": [
          "id",
          "anki_card_id",
          "deck_name",
          "sync_at"
        ],
        "title": "AnkiReviewQueueCardOut",
        "description": "Review-queue card (\u00a7T42 base shape) + per-card review metrics (T43).\n\n`retention` = lifetime true-retention (pass/total over non-learn reviews);\n`retrievability` = forgetting-curve recall estimate. Both None when the\ncard has no qualifying reviews / no scheduled interval. Data-only (V13)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}